feat: Always run callback from `handleShortcut`
If no shortcut, or Android version doesn't support shortcuts, the shortcut id
passed to the callback will be `undefined`.  This allows `handleShortcut()` to
be easily used to control the initial routing for the app.  Without this, we'd
have to wait for the callback, set some state variable somewhere, then defer
our usual routing.